requirements.txt
requests==2.27.1
lxml==3.8.0
mysql-connector-python==9.1.0
schedule==1.2.2